Level6 button hiding + save support
Game save now checks if you've "beaten" endless- gotten through a single loop of all 6 levels. If yes, show the Level6 button in the level select. If no, don't.

Also, added a new slot to the highScores array. Added support in SaveHandler.Load() to extend the array if it's too short- like if you've never loaded a V2 save before